added flow in the main and fixed mm

+++ mainMemory.h
@@ -1,1 +1,1 @@
-int fetchMainMemory(long la, long pdpa, struct Hardware *hardware, struct Process *process);
+short fetchMainMemory(long la, long pdpa, struct Hardware *hardware, struct Process *process);
